backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.7.0
python-decouple==3.8
gunicorn==23.0.0
psycopg2-binary==2.9.12
dj-database-url==2.3.0
whitenoise==6.9.0
drf-spectacular==0.28.0